move a bunch of test_schedule to null
